separated ChatBox & Report into separate components
